+++ twitcher/client.py
@@ -1,100 +1,113 @@
-import ssl
-from datetime import datetime
+import os
 
-import xmlrpc.client as xmlrpclib
-from urllib import parse as urlparse
+import requests
+from oauthlib.oauth2 import BackendApplicationClient
+from requests_oauthlib import OAuth2Session
+
+from twitcher.oauth2 import TOKEN_ENDPOINT, CLIENT_APP_ENDPOINT
 
 import logging
 LOGGER = logging.getLogger("TWITCHER")
 
-
-def _create_https_context(verify=True):
-    context = ssl._create_default_https_context()
-    if verify is False:
-        context.check_hostname = False
-        context.verify_mode = ssl.CERT_NONE
-    return context
-
-
-def _create_server(url, username=None, password=None, verify=True):
-    # TODO: disable basicauth when username is not set
-    username = username or 'nouser'
-    password = password or 'nopass'
-
-    parsed = urlparse.urlparse(url)
-    url = "%s://%s:%s@%s%s" % (parsed.scheme, username, password, parsed.netloc, parsed.path)
-    context = _create_https_context(verify=verify)
-    server = xmlrpclib.ServerProxy(url, context=context)
-    return server
-
-
-def xmlrpc_error_handler(wrapped):
-    def _handle_error(*args, **kwargs):
-        try:
-            result = wrapped(*args, **kwargs)
-        except xmlrpclib.Fault as e:
-            LOGGER.error("A fault occurred. {}".format(e))
-            raise
-        except xmlrpclib.ProtocolError as e:
-            LOGGER.error("A protocol error occurred. {}".format(e))
-            raise
-        except xmlrpclib.ResponseError as e:
-            LOGGER.error(
-                "A response error occured. Maybe service needs authentication with username and password? {}".format(e))
-            raise
-        except Exception as e:
-            LOGGER.error(
-                " Unknown error occured. "
-                "Maybe you need to use the \"--insecure\" option to access the service on HTTPS? "
-                "Is your service running and did you specify the correct service url (port)? "
-                "{}".format(e))
-            raise
-        else:
-            return result
-    return _handle_error
+KEYCLOAK_TOKEN_ENDPOINT = '/auth/realms/demo/protocol/openid-connect/token'
 
 
 class TwitcherService(object):
+    """TwitcherService is a twitcher client to talk to the twitcher service API."""
     def __init__(self, url, username=None, password=None, verify=True):
-        self.server = _create_server(url, username=username, password=password, verify=verify)
+        self.base_url = url
+        self.username = username
+        self.password = password
+        self.verify = verify
 
-    # tokens
+    def add_client_app(self, name=None, redirect_uri=None):
+        """Add a client application to twitcher with optional name."""
+        name = name or ''
+        redirect_uri = redirect_uri or ''
+        req_url = "{}{}?name={}&redirect_uri={}".format(
+            self.base_url,
+            CLIENT_APP_ENDPOINT,
+            name,
+            redirect_uri)
+        resp = requests.get(req_url, auth=(self.username, self.password), verify=self.verify)
+        if not resp.ok:
+            LOGGER.error("Could not add client app.")
+        else:
+            return resp.json()
 
-    @xmlrpc_error_handler
-    def generate_token(self, valid_in_hours=1):
-        return self.server.generate_token(valid_in_hours)
+    def fetch_token(self, client_id, client_secret, scope=None, keycloak=False):
+        """Get an access token with given scope."""
+        scope = scope or 'compute'
+        os.environ['OAUTHLIB_INSECURE_TRANSPORT'] = '1'
+        if keycloak is True:
+            endpoint = KEYCLOAK_TOKEN_ENDPOINT
+        else:
+            endpoint = TOKEN_ENDPOINT
+        token_url = "{}{}".format(self.base_url, endpoint)
+        client = BackendApplicationClient(client_id=client_id)
+        oauth = OAuth2Session(client=client)
+        return oauth.fetch_token(token_url,
+                                 scope=scope,
+                                 client_id=client_id,
+                                 client_secret=client_secret,
+                                 include_client_id=True,
+                                 verify=self.verify)
 
-    @xmlrpc_error_handler
-    def revoke_token(self, token):
-        return self.server.revoke_token(token)
-
-    @xmlrpc_error_handler
-    def revoke_all_tokens(self):
-        return self.server.revoke_all_tokens()
-
-    # service registry
-
-    @xmlrpc_error_handler
-    def register_service(self, name, url, data=None):
-        data = data or {}
-        return self.server.register_service(name, url, data)
-
-    @xmlrpc_error_handler
-    def unregister_service(self, name):
-        return self.server.unregister_service(name)
-
-    @xmlrpc_error_handler
     def list_services(self):
-        return self.server.list_services()
+        """List all registered OWS services."""
+        req_url = "{}/services".format(self.base_url)
+        resp = requests.get(req_url, auth=(self.username, self.password), verify=self.verify)
+        if not resp.ok:
+            LOGGER.error("Could not list services")
+        else:
+            return resp.json()
 
-    @xmlrpc_error_handler
+    def register_service(self, name, url, data=None):
+        """Register a service."""
+        data = data or {}
+        data.update({"name": name, "url": url})
+        req_url = "{}/services".format(self.base_url)
+        import json
+        resp = requests.post(req_url, data=json.dumps(data),
+                             auth=(self.username, self.password),
+                             headers={},
+                             verify=self.verify)
+        if not resp.ok:
+            LOGGER.error("Could not add service")
+        else:
+            return resp.json()
+
     def clear_services(self):
-        return self.server.clear_services()
+        """Remove all OWS services."""
+        req_url = "{}/services".format(self.base_url)
+        resp = requests.delete(req_url,
+                               auth=(self.username, self.password),
+                               headers={},
+                               verify=self.verify)
+        if not resp.ok:
+            LOGGER.error("Could not clear services")
+        else:
+            return resp.json()
 
-    @xmlrpc_error_handler
-    def get_service_by_url(self, url):
-        return self.server.get_service_by_url(url)
+    def get_service(self, name):
+        """Get an OWS service with given name."""
+        req_url = "{}/services/{}".format(self.base_url, name)
+        resp = requests.get(req_url,
+                            auth=(self.username, self.password),
+                            verify=self.verify)
+        if not resp.ok:
+            LOGGER.error("Could not get service")
+        else:
+            return resp.json()
 
-    @xmlrpc_error_handler
-    def get_service_by_name(self, name):
-        return self.server.get_service_by_name(name)
+    def unregister_service(self, name):
+        """Remove registered service with given name."""
+        req_url = "{}/services/{}".format(self.base_url, name)
+        resp = requests.delete(req_url,
+                               auth=(self.username, self.password),
+                               headers={},
+                               verify=self.verify)
+        if not resp.ok:
+            LOGGER.error("Could not remove service")
+        else:
+            return resp.json()