+++ twitcher/owsproxy.py
@@ -1,20 +1,22 @@
 """
-The owsproxy is based on `papyrus_ogcproxy <https://github.com/elemoine/papyrus_ogcproxy>`_
+The owsproxy is a service which acts as a proxy for registered OWS services.
+
+The implementation of owsproxy is based on `papyrus_ogcproxy <https://github.com/elemoine/papyrus_ogcproxy>`_
 
 See also: https://github.com/nive/outpost/blob/master/outpost/proxy.py
 """
-
-from urllib import parse as urlparse
-
 import requests
 
 from pyramid.response import Response
 from pyramid.settings import asbool
 from typing import TYPE_CHECKING
 
-from twitcher.adapter import get_adapter_factory
-from twitcher.owsexceptions import OWSAccessForbidden, OWSAccessFailed
-from twitcher.utils import replace_caps_url, get_settings, get_twitcher_url
+from twitcher.owsexceptions import OWSAccessForbidden, OWSAccessFailed, OWSException, OWSNoApplicableCode
+from twitcher.utils import (
+    replace_caps_url,
+    get_settings,
+    get_twitcher_url,
+    is_valid_url)
 
 import logging
 LOGGER = logging.getLogger('TWITCHER')
@@ -69,18 +71,19 @@ def _send_request(request, service, extra_path=None, request_params=None):
         url += '/' + extra_path
     if request_params:
         url += '?' + request_params
-    LOGGER.debug('url = %s', url)
+    LOGGER.debug('url = {}'.format(url))
 
     # forward request to target (without Host Header)
     h = dict(request.headers)
     h.pop("Host", h)
     h['Accept-Encoding'] = None
     #
-    service_type = service['type']
+    service_type = service.get('type', 'wps')
+    service_verify = service.get('verify', True)
     if service_type and (service_type.lower() != 'wps'):
         try:
             resp_iter = requests.request(method=request.method.upper(), url=url, data=request.body, headers=h,
-                                         stream=True, verify=service.verify)
+                                         stream=True, verify=service_verify)
         except Exception as e:
             return OWSAccessFailed("Request failed: {}".format(e))
 
@@ -92,7 +95,7 @@ def _send_request(request, service, extra_path=None, request_params=None):
     else:
         try:
             resp = requests.request(method=request.method.upper(), url=url, data=request.body, headers=h,
-                                    verify=service.verify)
+                                    verify=service_verify)
         except Exception as e:
             return OWSAccessFailed("Request failed: {}".format(e))
 
@@ -113,18 +116,18 @@ def _send_request(request, service, extra_path=None, request_params=None):
                 return OWSAccessForbidden(msg)
         else:
             # return OWSAccessFailed("Could not get content type from response.")
-            LOGGER.warn("Could not get content type from response")
+            LOGGER.warning("Could not get content type from response")
 
         try:
             if ct in ['text/xml', 'application/xml', 'text/xml;charset=ISO-8859-1']:
                 # replace urls in xml content
                 # ... if public URL is not configured use proxy url.
-                if service.has_purl():
-                    public_url = service.get('purl')
+                if is_valid_url(service.get('purl')):
+                    public_url = service['purl']
                 else:
                     public_url = request.route_url('owsproxy', service_name=service['name'])
                 # TODO: where do i need to replace urls?
-                content = replace_caps_url(resp.content, public_url, service.get('url'))
+                content = replace_caps_url(resp.content, public_url, service['url'])
             else:
                 # raw content
                 content = resp.content
@@ -151,41 +154,22 @@ def owsproxy_base_url(container):
 
 
 def owsproxy_view(request):
-    """
-    TODO: use ows exceptions
-    """
+    service_name = request.matchdict.get('service_name')
     try:
-        service_name = request.matchdict.get('service_name')
         extra_path = request.matchdict.get('extra_path')
-        adapter = get_adapter_factory(request)
-        store = adapter.servicestore_factory(request)
-        service = store.fetch_by_name(service_name)
-    except Exception as err:
-        # TODO: Store impl should raise appropriate exception like not authorized
-        return OWSAccessFailed("Could not find service {0} : {1}.".format(service_name, err))
-    else:
+        service = request.owsregistry.get_service_by_name(service_name)
+    except Exception:
+        return OWSAccessFailed("Could not find service: {}".format(service_name))
+    try:
+        if not request.is_verified:
+            raise OWSAccessForbidden("Access to service is forbidden.")
         return _send_request(request, service, extra_path, request_params=request.query_string)
-
-
-def owsproxy_delegate_view(request):
-    """
-    Delegates owsproxy request to external twitcher service.
-    """
-    twitcher_url = request.registry.settings.get('twitcher.url')
-    protected_path = request.registry.settings.get('twitcher.ows_proxy_protected_path', '/ows')
-    url = twitcher_url + protected_path + '/proxy'
-    if request.matchdict.get('service_name'):
-        url += '/' + request.matchdict.get('service_name')
-        if request.matchdict.get('access_token'):
-            url += '/' + request.matchdict.get('service_name')
-    url += '?' + urlparse.urlencode(request.params)
-    LOGGER.debug("delegate to owsproxy: %s", url)
-    # forward request to target (without Host Header)
-    # h = dict(request.headers)
-    # h.pop("Host", h)
-    resp = requests.request(method=request.method.upper(), url=url, data=request.body,
-                            headers=request.headers, verify=False)
-    return Response(resp.content, status=resp.status_code, headers=resp.headers)
+    except OWSException as exc:
+        LOGGER.warning("Security check failed but was not handled as expected by 'is_verified' method.", exc_info=exc)
+        raise
+    except Exception as exc:
+        LOGGER.exception("Security check failed due to unhandled error.", exc_info=exc)
+        raise OWSNoApplicableCode("Unhandled error: {!s}".format(exc))
 
 
 def owsproxy_defaultconfig(config):
@@ -193,21 +177,14 @@ def owsproxy_defaultconfig(config):
     settings = get_settings(config)
     if asbool(settings.get('twitcher.ows_proxy', True)):
         protected_path = owsproxy_base_path(settings)
-        LOGGER.debug('Twitcher {}/proxy enabled.'.format(protected_path))
 
+        config.include('twitcher.oauth2')
+        config.include('twitcher.owsregistry')
+        config.include('twitcher.owssecurity')
         config.add_route('owsproxy', protected_path + '/proxy/{service_name}')
         config.add_route('owsproxy_extra', protected_path + '/proxy/{service_name}/{extra_path:.*}')
-        config.add_route('owsproxy_secured', protected_path + '/proxy/{service_name}/{access_token}')
-
-        # use delegation mode?
-        if asbool(settings.get('twitcher.ows_proxy_delegate', False)):
-            LOGGER.debug('Twitcher {}/proxy delegation mode enabled.'.format(protected_path))
-            config.add_view(owsproxy_delegate_view, route_name='owsproxy')
-            config.add_view(owsproxy_delegate_view, route_name='owsproxy_secured')
-        else:
-            config.add_view(owsproxy_view, route_name='owsproxy')
-            config.add_view(owsproxy_view, route_name='owsproxy_secured')
-            config.add_view(owsproxy_view, route_name='owsproxy_extra')
+        config.add_view(owsproxy_view, route_name='owsproxy')
+        config.add_view(owsproxy_view, route_name='owsproxy_extra')
 
 
 def includeme(config):